requirements.txt
pyTelegramBotAPI==4.26.0
flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
pg8000==1.31.2
python-dotenv==1.0.0
